+++ src/model/usuariosmodel.ts
@@ -1,14 +1,14 @@
 import { Type } from "@fastify/type-provider-typebox";
 import { Static } from "@fastify/type-provider-typebox";
 export const Usuario = Type.Object(
-  {
+    {
     id_usuario: Type.Integer(),
     nombre: Type.String({ minLength: 2 }),
     isAdmin: Type.Boolean(),
-  },
-  {
+    },
+    {
     title: "Esquema para el Usuario",
-  }
+    }
 );
 
 export type Usuario = Static<typeof Usuario>;
